requirements.txt
python-dotenv~=1.0.0
aiogram~=2.25.1
barsdiary~=0.1.5
pymongo~=4.4.1
motor~=3.2.0
yookassa~=2.4.0